requirements.txt
python-box
pandas
numpy
seaborn
matplotlib
pyyaml~=5.3.1
python-dotenv~=0.13.0
yaml~=0.1.7
click~=7.1.1